backend/requirements.txt
Django>=4.2
django-cors-headers>=4.3
mercadopago>=2.2.0
python-dotenv>=1.0.0
